bugfix: do not always follow last link in path
when resolving path we don't want always
to resolve also the last components if it is a link. Cause in some cases
we want exactly the path of that link and not the one it points to.

lstat and readlink are some example of functions that don't want to
resolve the last link.